Add doc about proposed triple-purpose mining.
The new document describes the issues with pure merge- or stand-alone
mining and our proposed scheme to solve them.

It also describes the technical details of how blocks with PoW will look
like in Xyon to avoid the nVersion BIP 9 issue that Namecoin has, as
well as what RPC methods are provided to tools.