Add docs page about the proposed daemon interface for games.
We propose additional interfaces of the core daemon, that are tailored
specifically towards games and our game model.  This document describes
a first draft of what we think is a good interface that will make life
as easy as possible for game engine developers and allows them to focus
on the specific game, rather than things needed commonly for all games
on Xyon.