Add remark about new names and game states.
Where discussing that game engines must only depend on name operations
explicitly referencing the game, this adds another paragraph that
explicitly mentions that this applies to new names as well.  In other
words, game engines should never depend on the "list of all names", and
names should explicitly "join" a game when they want that.