Require nonce of actual blocks to be zero in Xyon.
Since we attach PoW to blocks as a separate data structure, the nonce in
a Xyon block header is not needed.  Explicitly state that it is required
to be zero in the mining doc.  By enforcing this requirement, we make
sure that the data that will be put into the blockchain is "controlled"
and that we can, if necessary, later repurpose the nonce easily.