Clarify the situation of new names in games.
This clarifies the rule about how game engines can and cannot treat
newly registered names.  It is not necessary for a particular name
p/alice to *itself* join the game -- it could also be that some existing
name p/bob mentions p/alice in a move.  (Although it is still expected
to be the typical case that p/alice joins on her own.)

The main rule is the bold paragraph above the change:  Game engines
should only depend on data from name updates referencing the game.

A situation in which this clarification is important is for currencies
implemented as "games":  There someone may well send coins to someone
else who did not yet mention the "game" themselves.